+++ src/Fantomas.Core.Tests/CompilerDirectivesTests.fs
@@ -676,7 +676,7 @@ type FunctionComponent =
         let elemType =
             ReactBindings.React.``lazy`` (fun () ->
                 // React.lazy requires a default export
-                (importValueDynamic f).``then`` (fun x -> createObj [ "default" ==> x ]))
+                (importValueDynamic f).``then``(fun x -> createObj [ "default" ==> x ]))
 
         fun props ->
             ReactElementType.create
@@ -868,7 +868,7 @@ type FunctionComponent =
         let elemType =
             ReactBindings.React.``lazy`` (fun () ->
                 // React.lazy requires a default export
-                (importValueDynamic f).``then`` (fun x -> createObj [ "default" ==> x ]))
+                (importValueDynamic f).``then``(fun x -> createObj [ "default" ==> x ]))
 
         fun props ->
             ReactElementType.create