+++ src/Fantomas.Core.Tests/PatternMatchingTests.fs
@@ -2390,7 +2390,7 @@ match
                                 -> AsyncSeq<_>) option
                 |} option
         |}
-      > (
+      >(
         source'
     ))
         .__proto__