+++ src/Fantomas.Core.Tests/Stroustrup/ExperimentalElmishTests.fs
@@ -850,7 +850,7 @@ let private useLocationDetail (auth0 : Auth0Hook) (roles : RolesHook) id =
 let private useLocationDetail (auth0 : Auth0Hook) (roles : RolesHook) id =
     let id = Guid.Parse(id)
     let eventCtx = React.useContext (eventContext)
-    let (creatorName, setCreatorName) = React.useState<string option> (None)
+    let (creatorName, setCreatorName) = React.useState<string option>(None)
 
     let location =
         React.useMemo ((fun () -> getLocation eventCtx.Events id), [| eventCtx.Events; id |])